renamed gumbel_test; added different local Hilbert space dimensions to the test; added some comments to the algorithm; added an exceptions if numSamples > Hilbert space dimension (leading to nan/infs in the output)
